pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arbiter-ai"
version = "0.1.0"
description = "Native PydanticAI evaluation with automatic cost tracking"
authors = [{name = "Evan Volgas", email = "evan.volgas@gmail.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11,<3.15"
keywords = ["ai", "llm", "evaluation", "testing", "metrics", "pydantic-ai", "cost-tracking"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
# Core dependencies - minimal for basic functionality
dependencies = [
    "pydantic>=2.12.0",
    "pydantic-ai>=1.14.0",  # Latest PydanticAI with current API
    "httpx>=0.28.0",
    "openai>=2.0.0",  # Latest OpenAI SDK
    "python-dotenv>=1.0.1",
    "aiofiles>=25.0.0",
    "logfire>=4.0.0",
    "numpy>=2.0.0",  # For vector operations
    "typing-extensions>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/ashita-ai/arbiter"
Documentation = "https://github.com/ashita-ai/arbiter#readme"
Repository = "https://github.com/ashita-ai/arbiter"
Issues = "https://github.com/ashita-ai/arbiter/issues"
Changelog = "https://github.com/ashita-ai/arbiter/releases"

[project.optional-dependencies]
# Model providers
anthropic = ["anthropic>=0.72.0"]
gemini = ["google-generativeai>=0.8.5"]
mistral = ["mistralai>=1.0.0"]
cohere = ["cohere>=5.0.0"]

# Scale features - optional fast backends
scale = [
    "faiss-cpu>=1.7.4",  # Fast in-memory vector similarity (CPU-only, no GPU required)
    "sentence-transformers>=2.2.0",  # Local embeddings (free, offline, deterministic)
    "pymilvus>=2.6.0",  # Optional Milvus for v2.0+ (enterprise scale)
]

# Storage backends
postgres = ["asyncpg>=0.29.0", "alembic>=1.13.0", "sqlalchemy>=2.0.0", "psycopg2-binary>=2.9.0"]
redis = ["redis>=5.0.0"]
storage = ["arbiter[postgres,redis]"]  # All storage backends

# Development tools
dev = [
    # Testing
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.15.0",
    "pytest-xdist>=3.8.0",  # Parallel test execution

    # Code quality
    "black>=25.0.0",
    "mypy>=1.18.0",
    "ruff>=0.14.0",
    "types-redis>=4.6.0",
    "types-aiofiles>=25.0.0",

    # Interactive development
    "rich>=14.0.0",

    # Documentation
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=0.30.0",

    # Development utilities
    "pre-commit>=4.0.0",
    "watchdog>=6.0.0",
    "psutil>=7.0.0",
    "setuptools>=80.0.0",
]

# All features
all = ["arbiter[anthropic,gemini,mistral,cohere,scale,redis,dev]"]

[tool.setuptools.packages.find]
where = ["."]
include = ["arbiter_ai*"]

[tool.setuptools.package-data]
arbiter_ai = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py311', 'py312', 'py313', 'py314']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
]

ignore = [
    "E501",   # line too long (handled by formatter)
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["arbiter_ai"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["--strict-markers", "--cov=arbiter_ai", "--cov-report=term-missing"]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "performance: marks tests as performance tests"
]

[tool.coverage.run]
omit = [
    "arbiter_ai/storage/__init__.py",
    "arbiter_ai/tools/__init__.py",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_reexport = true
strict_optional = true
strict_equality = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["pymilvus", "pymilvus.*", "logfire", "faiss", "sentence_transformers", "sentence_transformers.*"]
ignore_missing_imports = true
